Refactor Avatar component
