+++ packages/components/src/Avatar/Avatar.tsx
@@ -3,17 +3,26 @@ import glamorous, { CSSProperties, Div } from "glamorous"
 import { Theme, expandColor } from "@operational/theme"
 import { darken, readableTextColor } from "@operational/utils"
 
-import { WithTheme, Css, CssStatic } from "./types"
+import { WithTheme, Css, CssStatic } from "../types"
 
 export interface Props {
+  /** Name of the person */
   name: string
+  /** Title of the person */
   title?: string
+  /** Optionally display the full name next to the avatar */
   showName?: boolean
+  /** Hide initials if is true */
   hideInitials?: boolean
+  /** A URL to an image of the person */
   photo?: string
+  /** Glamorous CSS */
   css?: Css
+  /** Class name */
   className?: string
+  /** Color assigned to the avatar circle */
   color?: string
+  /** Automatically assign a deterministic color. (Invalidates `color` assignment)  */
   assignColor?: boolean
 }
